Fix ImagePopup text
