+++ src/components/Card/Card.js
@@ -1,8 +1,20 @@
 // import { CreateUserContext } from '../contexts/CurrentUserContext';
 // import { useContext } from 'react';
+import { useState } from 'react';
 import './card.css';
+import ImagePopup from '../ImagePopup/ImagePopup';
 
 function Card(props) {
+  const [isImagePopupOpen, setIsImagePopupOpen] = useState(false);
+
+  function handleCardClick(card) {
+    setIsImagePopupOpen(true);
+    console.log(card);
+  }
+
+  function closeImagePopup() {
+    setIsImagePopupOpen(false);
+  }
   // const currentUser = useContext(CreateUserContext);
   // const isOwn = props.card.owner === currentUser._id;
 
@@ -18,9 +30,9 @@ function Card(props) {
 
   // const cardLikesClassName = `card__likes ${isLiked && 'card__likes_active'}`;
 
-  function handleClick() {
-    props.onCardClick(props.card);
-  }
+  // function handleClick() {
+  //   props.onCardClick(props.card);
+  // }
 
   // function handleLikeClick() {
   //   props.onCardLike(props.card);
@@ -42,11 +54,17 @@ function Card(props) {
           src={props.card.url}
           className='card__image'
           alt={props.card.breeds[0].name}
-          onClick={handleClick}
+          onClick={handleCardClick}
         />
         <div className='card__title-container'>
           <h2 className='card__title'>{props.card.breeds[0].name}</h2>
           <p className='card__text'>{props.card.breeds[0].temperament}</p>
+          <ImagePopup
+            isOpen={isImagePopupOpen}
+            imgSrc={props.card.url}
+            imgName={props.card.breeds[0].name}
+            onClose={closeImagePopup}
+          />
           {/* <div className='card__like-container'>
             <button
               className={cardLikeButtonClassName}
