+++ src/components/ImagePopup/imagePopup.css
@@ -54,7 +54,7 @@
   }
 }
 
-.popup__info {
+.popup__text {
   font-size: 12px;
   font-weight: 400;
   line-height: 1.21;
